Add offline function to get data from net file.
This commit adds the ability to retrieve highway data from a net file
offline, in addition to the existing functionality of retrieving data
from OpenStreetMap. The new method reads the net file and creates a
shapefile containing the highway data.

Also changed some parameters in the main.py to get the ability to read
with offline connection.